CPU Optimisation - not much
- Call less often QDateTime::currentMSecsSinceEpoch() - not a lot of
optimisation here (1% of CPU time)
- We no not need to call qToBigEndian for the stored time as we received
exactly what we send. Probably not much optimisation either
- Performace analysis (linux perf) shows that 65% of the time is spend
in sendto, 10% in poll and 15% in recv, so I don't think I can do much
here, the only possibility I see to improve things is to allow multiple
threads for sending (and maybe buy a better NIC as the rtl8169... ).